:recycle: Modularize CPAC.utils.interfaces.function
