+++ CPAC/seg_preproc/__init__.py
@@ -12,12 +12,10 @@ from .utils import check_if_file_is_empty, \
                    pick_wm_class_2, \
                    mask_erosion, \
                    erosion, \
-                   hardcoded_antsJointLabelFusion, \
-                   pick_tissue_from_labels_file
+                   hardcoded_antsJointLabelFusion
 
 # List all functions
-__all__ = ['create_seg_preproc',
-           'process_segment_map',
+__all__ = ['process_segment_map',
            'check_if_file_is_empty',
            'pick_wm_prob_0',
            'pick_wm_prob_1',
@@ -28,7 +26,5 @@ __all__ = ['create_seg_preproc',
            'mask_erosion',
            'erosion',
            'hardcoded_antsJointLabelFusion',
-           'pick_tissue_from_labels_file',
-           'create_seg_preproc_template_based',
            'tissue_mask_template_to_t1',
            'create_seg_preproc_antsJointLabel_method']
